refactor: optimize expiry type and simplify permission model
  - Change expiry from uint256 to uint48 for storage efficiency
  - Replace onlyOwner with onlySelf modifier for cleaner access control
  - Remove redundant validator management functions from interface
  - Enhance test coverage with proper permission and initialization tests

+++ src/Types.sol
@@ -10,7 +10,7 @@ struct Call {
 struct BatchedCall {
     Call[] calls;
     uint256 nonce;
-    uint256 expiry;
+    uint48 expiry;
 }
 
 struct InitialOwner {
